+++ lib/config/lint_test.go
@@ -4,8 +4,6 @@ import (
 	"reflect"
 	"testing"
 
-	"gopkg.in/yaml.v3"
-
 	"github.com/Jeffail/benthos/v3/lib/config"
 	_ "github.com/Jeffail/benthos/v3/public/components/all"
 )
@@ -23,14 +21,14 @@ func TestConfigLints(t *testing.T) {
 		{
 			name:  "empty object",
 			conf:  `{}`,
-			lints: []string{},
+			lints: nil,
 		},
 		{
 			name: "root object type",
 			conf: `input:
   type: stdin
   kafka: {}`,
-			lints: []string{"line 3: path 'input': Key 'kafka' found but is ignored"},
+			lints: []string{"line 3: field kafka is invalid when the component type is stdin"},
 		},
 		{
 			name: "ignore tests section",
@@ -39,7 +37,7 @@ func TestConfigLints(t *testing.T) {
 tests:
   this: can just contain anything
   like_this: ["foo","bar"]`,
-			lints: []string{},
+			lints: nil,
 		},
 		{
 			name: "broker object type",
@@ -49,7 +47,19 @@ tests:
     inputs:
     - type: stdin
       kafka: {}`,
-			lints: []string{"line 6: path 'input.broker.inputs[0]': Key 'kafka' found but is ignored"},
+			lints: []string{"line 6: field kafka is invalid when the component type is stdin"},
+		},
+		{
+			name: "broker object type within resources",
+			conf: `resources:
+  inputs:
+    foo:
+      type: broker
+      broker:
+        inputs:
+          - type: stdin
+            kafka: {}`,
+			lints: []string{"line 8: field kafka is invalid when the component type is stdin"},
 		},
 		{
 			name: "broker object multiple types",
@@ -65,8 +75,8 @@ tests:
     - type: stdin
       stdin: {}`,
 			lints: []string{
-				"line 6: path 'input.broker.inputs[0]': Key 'kafka' found but is ignored",
-				"line 9: path 'input.broker.inputs[1]': Key 'stdin' found but is ignored",
+				"line 6: field kafka is invalid when the component type is stdin",
+				"line 8: field stdin is invalid when the component type is amqp",
 			},
 		},
 		{
@@ -80,31 +90,14 @@ tests:
         thisismadeup: true
         multipart: true`,
 			lints: []string{
-				"line 7: path 'input.broker.inputs[0].stdin': Key 'thisismadeup' found but is ignored",
-			},
-		},
-		{
-			name: "batch processor outside of input",
-			conf: `input:
-  type: amqp
-  amqp: {}
-pipeline:
-  processors:
-  - type: batch
-    batch: {}`,
-			lints: []string{
-				"line 6: path 'pipeline.processors[0].type': Type 'batch' is unsafe outside of the 'input' section, for more information read https://benthos.dev/docs/configuration/batching",
+				"line 7: field thisismadeup not recognised",
 			},
 		},
 	}
 
 	for _, test := range tests {
 		t.Run(test.name, func(tt *testing.T) {
-			conf := config.New()
-			if err := yaml.Unmarshal([]byte(test.conf), &conf); err != nil {
-				tt.Fatal(err)
-			}
-			lints, err := config.Lint([]byte(test.conf), conf)
+			lints, err := config.Lint([]byte(test.conf), config.New())
 			if err != nil {
 				tt.Fatal(err)
 			}