+++ lib/processor/while.go
@@ -45,16 +45,16 @@ If following a loop execution the number of messages in a batch is reduced to ze
 				`errored()`,
 				`this.urls.unprocessed.length() > 0`,
 			).HasDefault(""),
-			docs.FieldDeprecated("condition").HasType(docs.FieldCondition).OmitWhen(func(v, _ interface{}) bool {
+			docs.FieldDeprecated("condition").HasType(docs.FieldCondition).OmitWhen(func(v, _ interface{}) (string, bool) {
 				defaultBytes, err := yaml.Marshal(condition.NewConfig())
 				if err != nil {
-					return false
+					return "", false
 				}
 				var iDefault interface{}
 				if err = yaml.Unmarshal(defaultBytes, &iDefault); err != nil {
-					return false
+					return "", false
 				}
-				return cmp.Equal(v, iDefault)
+				return "field condition is deprecated in favour of check", cmp.Equal(v, iDefault)
 			}),
 			docs.FieldCommon("processors", "A list of child processors to execute on each loop.").Array().HasType(docs.FieldProcessor),
 		},