+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file.
 - New (experimental) Apache Pulsar input and output.
 - Field `codec` added to the `socket` output.
 - New Bloblang method `map_each_key`.
+- General config linting improvements
 
 ### Fixed